Clean up hooks implementations
